backend-apikey/requirements.txt

fastapi
uvicorn
pydantic
python-dotenv
